refactor(web): rename voice-chat service directory
Moves src/lib/zero/voice-chat-candidate/ to src/lib/zero/voice-chat/
and rewrites the /lib/zero/voice-chat-candidate/ path segment inside
every import specifier across apps/web. Identifier names are left
untouched in this commit; they flip in a later commit so the history
is reviewable layer by layer.

Refs #10877
