+++ turbo/apps/web/app/api/cron/voice-chat-cleanup/route.ts
@@ -4,7 +4,7 @@ import { initServices } from "../../../../src/lib/init-services";
 import { logger } from "../../../../src/lib/shared/logger";
 import { env } from "../../../../src/env";
 import { voiceChatSessions } from "../../../../src/db/schema/voice-chat";
-import { triggerReasoning } from "../../../../src/lib/zero/voice-chat-candidate/trigger-reasoning";
+import { triggerReasoning } from "../../../../src/lib/zero/voice-chat/trigger-reasoning";
 
 export const maxDuration = 60;
