+++ turbo/apps/web/app/api/internal/callbacks/voice-chat-candidate/route.ts
@@ -4,8 +4,8 @@ import { initServices } from "../../../../../src/lib/init-services";
 import { verifyCallback } from "../../../../../src/lib/infra/callback";
 import { agentRuns } from "../../../../../src/db/schema/agent-run";
 import { getRunOutputText } from "../../../../../src/lib/infra/run/extract-run-output";
-import { completeVoiceChatCandidateTask } from "../../../../../src/lib/zero/voice-chat-candidate/task-service";
-import { triggerReasoning } from "../../../../../src/lib/zero/voice-chat-candidate/trigger-reasoning";
+import { completeVoiceChatCandidateTask } from "../../../../../src/lib/zero/voice-chat/task-service";
+import { triggerReasoning } from "../../../../../src/lib/zero/voice-chat/trigger-reasoning";
 import { publishUserSignal } from "../../../../../src/lib/infra/realtime/client";
 import { isNotFound } from "../../../../../src/lib/shared/errors";
 import type { VoiceChatCallbackPayload } from "../../../../../src/lib/infra/callback/callback-payloads";
