+++ turbo/apps/web/app/api/zero/voice-chat-candidate/[id]/route.ts
@@ -1,8 +1,8 @@
 import { NextResponse } from "next/server";
 import { getAuthContext } from "../../../../../src/lib/auth/get-auth-context";
 import { initServices } from "../../../../../src/lib/init-services";
-import { getVoiceChatCandidateSession } from "../../../../../src/lib/zero/voice-chat-candidate/session-service";
-import { buildTalkerPayload } from "../../../../../src/lib/zero/voice-chat-candidate/talker-instructions";
+import { getVoiceChatCandidateSession } from "../../../../../src/lib/zero/voice-chat/session-service";
+import { buildTalkerPayload } from "../../../../../src/lib/zero/voice-chat/talker-instructions";
 import {
   isVoiceChatEnabled,
   notFoundResponse,
