+++ turbo/apps/web/app/api/zero/voice-chat-candidate/[id]/trigger-reasoning/route.ts
@@ -1,8 +1,8 @@
 import { NextResponse, after } from "next/server";
 import { getAuthContext } from "../../../../../../src/lib/auth/get-auth-context";
 import { initServices } from "../../../../../../src/lib/init-services";
-import { getVoiceChatCandidateSession } from "../../../../../../src/lib/zero/voice-chat-candidate/session-service";
-import { triggerReasoning } from "../../../../../../src/lib/zero/voice-chat-candidate/trigger-reasoning";
+import { getVoiceChatCandidateSession } from "../../../../../../src/lib/zero/voice-chat/session-service";
+import { triggerReasoning } from "../../../../../../src/lib/zero/voice-chat/trigger-reasoning";
 import {
   isVoiceChatEnabled,
   notFoundResponse,
