+++ turbo/apps/web/src/lib/zero/voice-chat/openai-token.ts
