+++ turbo/apps/web/src/lib/zero/voice-chat/trigger-reasoning.ts
